Phase 8.3.8: Fix no-any-return type errors (-19 errors)
Added float() and np.asarray() casts to handle numpy operations that return Any.
Added cast() for DataFrame returns from pandas operations.

Changes:
- statistical.py: Added float() casts for scipy stats functions
- matrix_ops.py: Added np.asarray() for vcov calculation
- standard_errors/utils.py: Added np.asarray() for 4 functions
- newey_west.py, driscoll_kraay.py: Added float() casts for kernel weights
- gmm/estimator.py: Added np.asarray() and bool() casts (4 functions)
- core/panel_data.py: Added cast(pd.DataFrame, ...) for 4 methods

Progress: 205 → 186 errors (-19 errors, 9.3% reduction)
Cumulative: 233 → 186 errors (-47 errors, 20.2% total reduction)

+++ panelbox/standard_errors/driscoll_kraay.py
@@ -213,7 +213,7 @@ class DriscollKraayStandardErrors:
             if lag == 0:
                 return 1.0
             z = 6 * np.pi * lag / (self.max_lags + 1) / 5
-            return 3 / z**2 * (np.sin(z) / z - np.cos(z))
+            return float(3 / z**2 * (np.sin(z) / z - np.cos(z)))
 
         else:
             raise ValueError(f"Unknown kernel: {self.kernel}")
